LayoutTests/editing/smart-lists/inserting-space-after-unicode-hyphen-generates-dashed-list.html

<!DOCTYPE html><!-- webkit-test-runner [ SmartListsAvailable=true ] -->
<html>
<head>
<meta charset="utf-8">
<script src="../editing.js"></script>
<script src="../../resources/dump-as-markup.js"></script>
</head>
<body>
<div id="test" contenteditable></div>
<script>
Markup.description('Typing the Unicode hyphen character (U+2010) followed by a space at the start of two consecutive lines should auto-convert them into an unordered dashed list, with the hyphen character in the custom list-style-type marker.');

var test = document.getElementById('test');
test.focus();
[...'‐ Hello\n‐ World'].forEach(typeCharacterCommand);
Markup.dump('test');
</script>
</body>
</html>
